+++ src/Form/Field/UploadField.php
@@ -5,7 +5,6 @@ namespace Encore\Admin\Form\Field;
 use Encore\Admin\Form;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\URL;
-use Illuminate\Support\MessageBag;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 trait UploadField
@@ -39,6 +38,57 @@ trait UploadField
     protected $useUniqueName = false;
 
     /**
+     * If use sequence name to store upload file.
+     *
+     * @var bool
+     */
+    protected $useSequenceName = false;
+
+    /**
+     * Retain file when delete record from DB.
+     *
+     * @var bool
+     */
+    protected $retainable = false;
+
+    /**
+     * @var bool
+     */
+    protected $downloadable = true;
+
+    /**
+     * Configuration for setting up file actions for newly selected file thumbnails in the preview window.
+     *
+     * @var array
+     */
+    protected $fileActionSettings = [
+        'showRemove' => false,
+        'showDrag'   => false,
+    ];
+
+    /**
+     * Controls the storage permission. Could be 'private' or 'public'.
+     *
+     * @var string
+     */
+    protected $storagePermission;
+
+    /**
+     * @var array
+     */
+    protected $fileTypes = [
+        'image'  => '/^(gif|png|jpe?g|svg)$/i',
+        'html'   => '/^(htm|html)$/i',
+        'office' => '/^(docx?|xlsx?|pptx?|pps|potx?)$/i',
+        'gdocs'  => '/^(docx?|xlsx?|pptx?|pps|potx?|rtf|ods|odt|pages|ai|dxf|ttf|tiff?|wmf|e?ps)$/i',
+        'text'   => '/^(txt|md|csv|nfo|ini|json|php|js|css|ts|sql)$/i',
+        'video'  => '/^(og?|mp4|webm|mp?g|mov|3gp)$/i',
+        'audio'  => '/^(og?|mp3|mp?g|wav)$/i',
+        'pdf'    => '/^(pdf)$/i',
+        'flash'  => '/^(swf)$/i',
+    ];
+
+    /**
      * Initialize the storage instance.
      *
      * @return void.
@@ -55,26 +105,30 @@ trait UploadField
      */
     protected function setupDefaultOptions()
     {
-        $defaultOptions = [
+        $defaults = [
             'overwriteInitial'     => false,
             'initialPreviewAsData' => true,
             'browseLabel'          => trans('admin.browse'),
+            'cancelLabel'          => trans('admin.cancel'),
             'showRemove'           => false,
             'showUpload'           => false,
-            'initialCaption'       => $this->initialCaption($this->value),
+            'showCancel'           => false,
+            'dropZoneEnabled'      => false,
             'deleteExtraData'      => [
-                $this->column            => static::FILE_DELETE_FLAG,
-                static::FILE_DELETE_FLAG => '',
-                '_token'                 => csrf_token(),
-                '_method'                => 'PUT',
+                $this->formatName($this->column) => static::FILE_DELETE_FLAG,
+                static::FILE_DELETE_FLAG         => '',
+                '_token'                         => csrf_token(),
+                '_method'                        => 'PUT',
             ],
         ];
 
         if ($this->form instanceof Form) {
-            $defaultOptions['deleteUrl'] = $this->form->resource().'/'.$this->form->model()->getKey();
+            $defaults['deleteUrl'] = $this->form->resource().'/'.$this->form->model()->getKey();
         }
 
-        $this->options($defaultOptions);
+        $defaults = array_merge($defaults, ['fileActionSettings' => $this->fileActionSettings]);
+
+        $this->options($defaults);
     }
 
     /**
@@ -84,10 +138,75 @@ trait UploadField
      */
     protected function setupPreviewOptions()
     {
-        $this->options([
-            //'initialPreview'        => $this->preview(),
-            'initialPreviewConfig'  => $this->initialPreviewConfig(),
-        ]);
+        $initialPreviewConfig = $this->initialPreviewConfig();
+
+        $this->options(compact('initialPreviewConfig'));
+    }
+
+    /**
+     * @return array|bool
+     */
+    protected function guessPreviewType($file)
+    {
+        $filetype = 'other';
+        $ext = strtok(strtolower(pathinfo($file, PATHINFO_EXTENSION)), '?');
+
+        foreach ($this->fileTypes as $type => $pattern) {
+            if (preg_match($pattern, $ext) === 1) {
+                $filetype = $type;
+                break;
+            }
+        }
+
+        $extra = ['type' => $filetype];
+
+        if ($filetype == 'video') {
+            $extra['filetype'] = "video/{$ext}";
+        }
+
+        if ($this->downloadable) {
+            $extra['downloadUrl'] = $this->objectUrl($file);
+        }
+
+        return $extra;
+    }
+
+    /**
+     * Indicates if the underlying field is downloadable.
+     *
+     * @param bool $downloadable
+     *
+     * @return $this
+     */
+    public function downloadable($downloadable = true)
+    {
+        $this->downloadable = $downloadable;
+
+        return $this;
+    }
+
+    /**
+     * Allow use to remove file.
+     *
+     * @return $this
+     */
+    public function removable()
+    {
+        $this->fileActionSettings['showRemove'] = true;
+
+        return $this;
+    }
+
+    /**
+     * Indicates if the underlying field is retainable.
+     *
+     * @return $this
+     */
+    public function retainable($retainable = true)
+    {
+        $this->retainable = $retainable;
+
+        return $this;
     }
 
     /**
@@ -109,20 +228,26 @@ trait UploadField
      *
      * @param string $disk Disks defined in `config/filesystems.php`.
      *
+     * @throws \Exception
+     *
      * @return $this
      */
     public function disk($disk)
     {
-        if (!array_key_exists($disk, config('filesystems.disks'))) {
-            $error = new MessageBag([
-                'title'   => 'Config error.',
-                'message' => "Disk [$disk] not configured, please add a disk config in `config/filesystems.php`.",
-            ]);
+        try {
+            $this->storage = Storage::disk($disk);
+        } catch (\Exception $exception) {
+            if (!array_key_exists($disk, config('filesystems.disks'))) {
+                admin_error(
+                    'Config error.',
+                    "Disk [$disk] not configured, please add a disk config in `config/filesystems.php`."
+                );
 
-            return session()->flash('error', $error);
+                return $this;
+            }
+
+            throw $exception;
         }
-
-        $this->storage = Storage::disk($disk);
 
         return $this;
     }
@@ -189,6 +314,18 @@ trait UploadField
     }
 
     /**
+     * Use sequence name for store upload file.
+     *
+     * @return $this
+     */
+    public function sequenceName()
+    {
+        $this->useSequenceName = true;
+
+        return $this;
+    }
+
+    /**
      * Get store name of upload file.
      *
      * @param UploadedFile $file
@@ -201,10 +338,12 @@ trait UploadField
             return $this->generateUniqueName($file);
         }
 
-        if (is_callable($this->name)) {
-            $callback = $this->name->bindTo($this);
+        if ($this->useSequenceName) {
+            return $this->generateSequenceName($file);
+        }
 
-            return call_user_func($callback, $file);
+        if ($this->name instanceof \Closure) {
+            return $this->name->call($this, $file);
         }
 
         if (is_string($this->name)) {
@@ -239,6 +378,10 @@ trait UploadField
     {
         $this->renameIfExists($file);
 
+        if (!is_null($this->storagePermission)) {
+            return $this->storage->putFileAs($this->getDirectory(), $file, $this->name, $this->storagePermission);
+        }
+
         return $this->storage->putFileAs($this->getDirectory(), $file, $this->name);
     }
 
@@ -269,6 +412,10 @@ trait UploadField
             return $path;
         }
 
+        if ($this->storage) {
+            return $this->storage->url($path);
+        }
+
         return Storage::disk(config('admin.upload.disk'))->url($path);
     }
 
@@ -281,7 +428,29 @@ trait UploadField
      */
     protected function generateUniqueName(UploadedFile $file)
     {
-        return md5(uniqid()).'.'.$file->guessExtension();
+        return md5(uniqid()).'.'.$file->getClientOriginalExtension();
+    }
+
+    /**
+     * Generate a sequence name for uploaded file.
+     *
+     * @param UploadedFile $file
+     *
+     * @return string
+     */
+    protected function generateSequenceName(UploadedFile $file)
+    {
+        $index = 1;
+        $extension = $file->getClientOriginalExtension();
+        $original = $file->getClientOriginalName();
+        $new = sprintf('%s_%s.%s', $original, $index, $extension);
+
+        while ($this->storage->exists("{$this->getDirectory()}/$new")) {
+            $index++;
+            $new = sprintf('%s_%s.%s', $original, $index, $extension);
+        }
+
+        return $new;
     }
 
     /**
@@ -291,8 +460,30 @@ trait UploadField
      */
     public function destroy()
     {
+        if ($this->retainable) {
+            return;
+        }
+
+        if (method_exists($this, 'destroyThumbnail')) {
+            $this->destroyThumbnail();
+        }
+
         if ($this->storage->exists($this->original)) {
             $this->storage->delete($this->original);
         }
+    }
+
+    /**
+     * Set file permission when stored into storage.
+     *
+     * @param string $permission
+     *
+     * @return $this
+     */
+    public function storagePermission($permission)
+    {
+        $this->storagePermission = $permission;
+
+        return $this;
     }
 }
